Add an earlyStartup of PyDev Scripting console
Create the PyDev Scripting as part of the Eclipse startup. This is done
as part of startup so that future creations of IPythonInterpreter (with
newPythonInterpreter) don't cause the PyDev Scripting console to pop-up
on top of the current interactive console.

A particular case that hits regularly is if the PyDev Scripting console
has not yet been created and you request an completion in a pydev
interactive console, the scripting console is then created, popping up
on top of the interactive console the completion was requested on.

This behaviour can easily be overridden by unchecking "Jython Plug-in" in
the "Startup and Shutdown" Preference page.